requirements.txt
pyservicebinding
sqlalchemy
pysqlite3
python-dotenv
feedparser
pytest
fastapi
uvicorn
pydantic
lru_cache
celery
requests
httpx
kubernetes
